Improve get_docstrings function
Take advantage of ast.walk to walk into the ast tree and grab all
functions node even if they are not inside a TestCase class. All
functions that name starts with "test_" will be considered as being a
test function or method and the docstring will be checked.

This also reduces the complexity of the function since we have reduced
the number of for loops and some indentation level.