Add context parse management
Add the concept of contexts when parsing the tests functions' docstring. Now
testimony tags can be added to the module and class level docstrings.

The context resolution will be done in the following order function > class >
module, if a tag is found then the resolution will stop.

Also start storing non recognized tags to improve Testimony usage as a library.

Close #93