Improve docstring parser to not generate a docinfo
Disable the DocInfo transformation. Without that transformation the
first field list element will not be transformed into a docinfo element.
This will normalize the parsed content when there is or there isn't any
element before the tokens.

Thanks @mbukatov for the sample code on how to disable the DocInfo
transformation.